+++ asserts/asserts_test.go
@@ -16,7 +16,6 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"strings"
 	"sync"
 	"testing"
 	"time"
@@ -194,35 +193,35 @@ func TestAssertRange(t *testing.T) {
 	failingAssert.Range(map[int]int{3: 1, 2: 2, 1: 3}, 5, 10, "map length out of range")
 }
 
-// TestAssertContents tests the Contents() assertion.
-func TestAssertContents(t *testing.T) {
+// TestAssertContains tests the Contains() and NotContains() assertion.
+func TestAssertContains(t *testing.T) {
 	successfulAssert := successfulAsserts(t)
 	failingAssert := failingAsserts(t)
 
-	successfulAssert.Contents("bar", "foobarbaz")
-	successfulAssert.Contents(4711, []int{1, 2, 3, 4711, 5, 6, 7, 8, 9})
-	failingAssert.Contents(4711, "12345-4711-67890")
-	failingAssert.Contents(4711, "foo")
-	failingAssert.Contents(4711, []interface{}{1, "2", 3, "4711", 5, 6, 7, 8, 9})
-	successfulAssert.Contents("4711", []interface{}{1, "2", 3, "4711", 5, 6, 7, 8, 9})
-	failingAssert.Contents("foobar", []byte("the quick brown fox jumps over the lazy dog"))
+	successfulAssert.Contains("bar", "foobarbaz")
+	successfulAssert.Contains(4711, []int{1, 2, 3, 4711, 5, 6, 7, 8, 9})
+	failingAssert.Contains(4711, "12345-4711-67890")
+	failingAssert.Contains(4711, "foo")
+	failingAssert.Contains(4711, []interface{}{1, "2", 3, "4711", 5, 6, 7, 8, 9})
+	successfulAssert.Contains("4711", []interface{}{1, "2", 3, "4711", 5, 6, 7, 8, 9})
+	failingAssert.Contains("foobar", []byte("the quick brown fox jumps over the lazy dog"))
 
-	successfulAssert.NotContents("yadda", "foobarbaz")
-	successfulAssert.NotContents(123, []int{1, 2, 3, 4711, 5, 6, 7, 8, 9})
-	failingAssert.NotContents("4711", "12345-4711-67890")
-	failingAssert.NotContents("oba", "foobar")
-	failingAssert.NotContents("4711", []interface{}{1, "2", 3, "4711", 5, 6, 7, 8, 9})
-	successfulAssert.NotContents(4711, []interface{}{1, "2", 3, "4711", 5, 6, 7, 8, 9})
-	failingAssert.NotContents("fox", []byte("the quick brown fox jumps over the lazy dog"))
+	successfulAssert.NotContains("yadda", "foobarbaz")
+	successfulAssert.NotContains(123, []int{1, 2, 3, 4711, 5, 6, 7, 8, 9})
+	failingAssert.NotContains("4711", "12345-4711-67890")
+	failingAssert.NotContains("oba", "foobar")
+	failingAssert.NotContains("4711", []interface{}{1, "2", 3, "4711", 5, 6, 7, 8, 9})
+	successfulAssert.NotContains(4711, []interface{}{1, "2", 3, "4711", 5, 6, 7, 8, 9})
+	failingAssert.NotContains("fox", []byte("the quick brown fox jumps over the lazy dog"))
 }
 
-// TestAssertContentsPrint test the visualization of failing content tests.
-func TestAssertContentsPrint(t *testing.T) {
+// TestAssertContainsPrint test the visualization of failing content tests.
+func TestAssertContainsPrint(t *testing.T) {
 	assert := asserts.NewTesting(t, asserts.NoFailing)
 
 	assert.Logf("printing of failing content tests")
-	assert.Contents("foobar", []byte("the quick brown fox jumps over the lazy dog"), "test fails but passes, just visualization")
-	assert.Contents([]byte("foobar"), []byte("the quick brown ..."), "test fails but passes, just visualization")
+	assert.Contains("foobar", []byte("the quick brown fox jumps over the lazy dog"), "test fails but passes, just visualization")
+	assert.Contains([]byte("foobar"), []byte("the quick brown ..."), "test fails but passes, just visualization")
 }
 
 // TestOffsetPrint test the correct visualization when printing
@@ -568,7 +567,7 @@ func TestValidationAssertion(t *testing.T) {
 	details := failures.Details()
 	location, fun := details[0].Location()
 	tt := details[0].Test()
-	if location != "asserts_test.go:553:0:" || fun != "TestValidationAssertion" {
+	if location != "asserts_test.go:554:0:" || fun != "TestValidationAssertion" {
 		t.Errorf("wrong location %q or function %q of first detail", location, fun)
 	}
 	if tt != asserts.True {
@@ -576,7 +575,7 @@ func TestValidationAssertion(t *testing.T) {
 	}
 	location, fun = details[1].Location()
 	tt = details[1].Test()
-	if location != "asserts_test.go:554:0:" || fun != "TestValidationAssertion" {
+	if location != "asserts_test.go:555:0:" || fun != "TestValidationAssertion" {
 		t.Errorf("wrong location %q or function %q of second detail", location, fun)
 	}
 	if tt != asserts.Equal {
@@ -618,9 +617,9 @@ func TestSetPrinter(t *testing.T) {
 
 	b := bp.Flush()
 	assert.Length(b, 3)
-	assert.Contents(b[0], "first")
-	assert.Contents(b[1], "second")
-	assert.Contents(b[2], "third")
+	assert.Contains(b[0], "first")
+	assert.Contains(b[1], "second")
+	assert.Contains(b[2], "third")
 	b = bp.Flush()
 	assert.Length(b, 0)
 }
@@ -647,14 +646,6 @@ func (f *metaFailer) Logf(format string, args ...interface{}) {
 }
 
 func (f *metaFailer) Fail(test asserts.Test, obtained, expected interface{}, msgs ...string) bool {
-	msg := strings.Join(msgs, " ")
-	if msg != "" {
-		msg = " [" + msg + "]"
-	}
-	format := "testing assert %q failed: '%v' (%v) <> '%v' (%v)" + msg
-	obtainedVD := asserts.ValueDescription(obtained)
-	expectedVD := asserts.ValueDescription(expected)
-	f.Logf(format, test, obtained, obtainedVD, expected, expectedVD)
 	if f.fail {
 		f.t.FailNow()
 	}